reduce dependency of session
