+++ ts/packages/dispatcher/src/context/appAgentStateConfig.ts
@@ -1,0 +1,10 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+export type AppAgentStateConfig = {
+    schemas: Record<string, boolean>;
+    actions: Record<string, boolean>;
+    commands: Record<string, boolean>;
+};
+
+export const appAgentStateKeys = ["schemas", "actions", "commands"] as const;
