+++ ts/packages/dispatcher/src/context/session.ts
@@ -11,7 +11,10 @@ import fs from "node:fs";
 import path from "node:path";
 import { getUniqueFileName, getYMDPrefix } from "../utils/fsUtils.js";
 import ExifReader from "exifreader";
-import { AppAgentStateConfig, appAgentStateKeys } from "./appAgentManager.js";
+import {
+    AppAgentStateConfig,
+    appAgentStateKeys,
+} from "./appAgentStateConfig.js";
 import {
     cloneConfig,
     isEmptySettings,
